imdb_service/requirements.txt
Flask-JWT
Flask-JWT-Extended
Flask-RESTful
Flask>=1.0
Werkzeug==0.16.1
alembic
bs4
cryptography
flasgger
flask-restful-swagger
flask_admin
flask_assets
flask_cors
flask_debugtoolbar
flask_login==0.4.1
flask_mail
flask_migrate
flask_restplus
flask_script
flask_sqlalchemy
flask_wtf
gevent
gunicorn==19.9.0
json2html
lxml
pymysql
python-ldap
pytz
redis
requests
flake8
urllib3
python-dotenv
pandas
numpy
flask-bcrypt
chardet
